+++ tests/test_account.py
@@ -138,3 +138,71 @@ def test_wrong_login(testclient, slapd_connection, user):
     res.form["password"] = "incorrect horse"
     res = res.form.submit(status=200)
     assert "The login 'invalid' does not exist" in res.text
+
+
+def test_admin_self_deletion(testclient, slapd_connection):
+    User.ocs_by_name(slapd_connection)
+    admin = User(
+        objectClass=["inetOrgPerson"],
+        cn="Temp admin",
+        sn="admin",
+        uid="temp",
+        mail="temp@temp.com",
+        userPassword="{SSHA}Vmgh2jkD0idX3eZHf8RzGos31oerjGiU",
+    )
+    admin.save(slapd_connection)
+    with testclient.session_transaction() as sess:
+        sess["user_dn"] = [admin.dn]
+
+    res = testclient.get("/profile/temp")
+    res = (
+        res.form.submit(name="action", value="delete", status=302)
+        .follow(status=302)
+        .follow(status=200)
+    )
+
+    testclient.app.config["SELF_DELETION"] = True
+
+    testclient.app.config["SELF_DELETION"] = False
+
+    with testclient.app.app_context():
+        assert User.get("temp", conn=slapd_connection) is None
+
+    with testclient.session_transaction() as sess:
+        assert not sess.get("user_dn")
+
+
+def test_user_self_deletion(testclient, slapd_connection):
+    User.ocs_by_name(slapd_connection)
+    user = User(
+        objectClass=["inetOrgPerson"],
+        cn="Temp user",
+        sn="user",
+        uid="temp",
+        mail="temp@temp.com",
+        userPassword="{SSHA}Vmgh2jkD0idX3eZHf8RzGos31oerjGiU",
+    )
+    user.save(slapd_connection)
+    with testclient.session_transaction() as sess:
+        sess["user_dn"] = [user.dn]
+
+    testclient.app.config["SELF_DELETION"] = False
+    res = testclient.get("/profile/temp")
+    assert "Delete my account" not in res
+
+    testclient.app.config["SELF_DELETION"] = True
+    res = testclient.get("/profile/temp")
+    assert "Delete my account" in res
+    res = (
+        res.form.submit(name="action", value="delete", status=302)
+        .follow(status=302)
+        .follow(status=200)
+    )
+
+    with testclient.app.app_context():
+        assert User.get("temp", conn=slapd_connection) is None
+
+    with testclient.session_transaction() as sess:
+        assert not sess.get("user_dn")
+
+    testclient.app.config["SELF_DELETION"] = False